requirements.txt
python-telegram-bot[webhooks]>=20.0,<21.0  # Telegram Bot API 和 Webhook 支持
asyncpg>=0.28.0,<0.30.0                   # 异步 PostgreSQL 驱动
pyyaml>=6.0,<7.0                          # YAML 文件解析
structlog==24.4.0                         # 结构化日志
pytz>=2024.1,<2025.0                      # 时区支持
aiofiles==23.2.1
aiohttp>=3.8.0